+++ tests/testthat/test-synteny.R
@@ -3,15 +3,22 @@
 data(synnet)
 
 #----Start tests----------------------------------------------------------------
+test_that("fit_sft() returns a numeric vector of R squared values", {
+
+    rsquared <- fit_sft(synnet)
+
+    expect_equal(class(rsquared), "numeric")
+})
+
 test_that("assess_synnet() works", {
     x <- assess_synnet(synnet)
 
     # Simulated synteny network without cols 'anchor1' and 'anchor2'
-    sim_synnet <- synnet
-    names(sim_synnet) <- c("name1", "name2")
+    sim_synnet <- matrix(c(1,2,3,4), ncol = 4)
 
-    expect_equal(ncol(x), 3)
-    expect_equal(names(x), c("CC", "Node_number", "Score"))
+
+    expect_equal(ncol(x), 4)
+    expect_equal(names(x), c("CC", "Node_count", "Rsquared", "Score"))
     expect_error(assess_synnet(sim_synnet))
 })
 
@@ -30,7 +37,7 @@ test_that("assess_synnet_list() returns a data frame", {
     x2 <- assess_synnet_list(synnet_list2)
 
     expect_equal(class(x), "data.frame")
-    expect_equal(ncol(x), 4)
+    expect_equal(ncol(x), 5)
     expect_true("Network" %in% names(x))
 
     expect_equal(class(x2), "data.frame")